requirements.txt
Flask>=3.0.0
Werkzeug>=3.0.0
python-dotenv>=1.0.1
filetype>=1.2.0

# Azure
azure-storage-blob>=12.20.0
azure-identity>=1.16.0
gunicorn>=21.0.0

# Image handling
Pillow>=10.0.0

# (Optional) HTTP utilities
requests>=2.31.0

# mangodb
pymongo>=4.6.0
bcrypt>=4.1.0

#emails
resend>=0.8.0